requirements.txt
aspose-words==23.1.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Faker==15.3.2
filelock==3.8.0
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
platformdirs==2.5.4
python-dateutil==2.8.2
python-docx==0.8.11
requests==2.28.2
six==1.16.0
urllib3==1.26.14
virtualenv==20.17.0
Werkzeug==2.2.2
